Edit random number generation and other things.
I started editing the random number generator for CORE-1197, but it
looks like the important things were already done by Nevers and
Keegan.  At any rate, what I have written here compiles with the
master branch of Reach (which includes a commit by Nevers adding
support for modulo on digests), and I believe conforms to the spec
provided for on-chain randomness.  However, I think it would be better
if we could get the digest of the previous block (which I don't know
how to do, I don't think Reach currently has a feature for that)
rather than just the round number and timestamp.  On Algorand the
block time is pretty consistent as far as I understand, so you could
predict both a block number and time combination pretty far in
advance.  As long as you have a period with nobody else playing the
game and affecting the stored random seed, you could have a long time
to find a matching value to provide to get the NFT you want.  If we
included a block digest then it would be effectively impossible to
start computing values to get what you want until the previous block
is published.

Anyway, after editing the line with the random number generation, I
was reading through tho code to try to understand what it was doing
and how it would use the random number generator.  I ultimately edited
a bunch of other things to fix some issues that I saw based on my
understanding.  These edits got mixed up with the random number
generator edits, so now I'm doing a mixed commit rather than nice
individual patches.

That said, I don't see a test file and I didn't write one, so what I
wrote could all be wrong, too.  But I hope it helps, at least.